bot_app/requirements.txt
aiogram<3.0
aiohttp==3.6.2
aioredis<2.0
python-dotenv